+++ crates/mimic_declare/src/node/entity.rs
@@ -76,11 +76,7 @@ impl HasSchemaPart for Entity {
 impl HasTraits for Entity {
     fn traits(&self) -> Vec<Trait> {
         let mut traits = self.traits.clone().with_type_traits();
-        traits.extend(vec![
-            Trait::EntityAccessor,
-            Trait::EntityKind,
-            Trait::EntityFixture,
-        ]);
+        traits.extend(vec![Trait::EntityKind, Trait::EntityFixture]);
 
         traits.list()
     }
@@ -91,7 +87,6 @@ impl HasTraits for Entity {
         match t {
             Trait::Default if self.fields.has_default() => DefaultTrait::tokens(self),
             Trait::From => FromTrait::tokens(self),
-            Trait::EntityAccessor => EntityAccessorTrait::tokens(self),
             Trait::EntityKind => EntityKindTrait::tokens(self),
             Trait::TypeView => TypeViewTrait::tokens(self),
             Trait::ValidateAuto => ValidateAutoTrait::tokens(self),